Only fetch JDK if needed

+++ jenkins/ec2_bootstrap.sh
@@ -346,7 +346,9 @@ install_packages() {
 }
 
 install_jdk() {
-    wget --continue --no-check-certificate -O /tmp/jdk8.tar.gz --header "Cookie: oraclelicense=a" 'https://edelivery.oracle.com/otn-pub/java/jdk/8u161-b12/2f38c3b165be4555a1fa6e98c45e0808/jdk-8u161-linux-x64.tar.gz'
+    if [[ ! -f /tmp/jdk8.tar.gz ]]; then
+      wget --continue --no-check-certificate -O /tmp/jdk8.tar.gz --header "Cookie: oraclelicense=a" 'https://edelivery.oracle.com/otn-pub/java/jdk/8u161-b12/2f38c3b165be4555a1fa6e98c45e0808/jdk-8u161-linux-x64.tar.gz'
+    fi
     cd /tmp
     tar xfz jdk8.tar.gz
     mv jdk1.8.0_161 /usr/local/java
